Refactor integrations::github to share types with other integrations
